# services/backend/docs/openapi/backend.yaml
openapi: 3.1.0
info:
  title: Gift Suggestion Backend API
  version: 0.2.0
  description: |
    Backend API gateway for the Gift Suggestion Web Service.
    The specification is maintained manually and currently covers health,
    auth/user foundation, catalog read endpoints, wishlist endpoints,
    admin catalog import endpoints, recommendation endpoints, tracking endpoints,
    and VK integration scaffold endpoints.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/users:
    post:
      tags: [auth, users]
      summary: Register user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterEnvelope'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Login user
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Access token issued, refresh token returned via HttpOnly cookie
          headers:
            Set-Cookie:
              description: Refresh token cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginEnvelope'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      summary: Refresh access token
      operationId: refreshToken
      parameters:
        - in: cookie
          name: refresh_token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: New access token issued, refresh cookie rotated
          headers:
            Set-Cookie:
              description: Rotated refresh token cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshEnvelope'
        '401':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /api/v1/auth/password-reset/request:
    post:
      tags: [auth]
      summary: Request password reset
      operationId: requestPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '202':
          description: Password reset request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedEnvelope'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/auth/password-reset/confirm:
    post:
      tags: [auth]
      summary: Confirm password reset
      operationId: confirmPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirmRequest'
      responses:
        '200':
          description: Password reset confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedEnvelope'
        '400':
          description: Validation error or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/auth/email-verification/confirm:
    post:
      tags: [auth]
      summary: Confirm email verification
      operationId: confirmEmailVerification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailVerificationConfirmRequest'
      responses:
        '200':
          description: Email verification confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedEnvelope'
        '400':
          description: Validation error or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    Meta:
      type: object
      properties:
        request_id:
          type: string
      additionalProperties: false
    ErrorPayload:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
      additionalProperties: false
    ErrorEnvelope:
      type: object
      required: [status, error, meta]
      properties:
        status:
          type: string
          const: error
        error:
          $ref: '#/components/schemas/ErrorPayload'
        meta:
          $ref: '#/components/schemas/Meta'
      additionalProperties: false
...
    UserProfile:
      type: object
      required: [id, email, email_verified, role, display_name, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        email_verified:
          type: boolean
        role:
          type: string
        display_name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_login_at:
          type: string
          format: date-time
      additionalProperties: false
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 72
        display_name:
          type: string
          maxLength: 120
      additionalProperties: false
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      additionalProperties: false
    PasswordResetRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
      additionalProperties: false
    PasswordResetConfirmRequest:
      type: object
      required: [token, new_password]
      properties:
        token:
          type: string
        new_password:
          type: string
          minLength: 8
          maxLength: 72
      additionalProperties: false
    EmailVerificationConfirmRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
      additionalProperties: false
...
    AuthPayload:
      type: object
      required: [access_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          format: int64
      additionalProperties: false
    AcceptedPayload:
      type: object
      required: [accepted]
      properties:
        accepted:
          type: boolean
      additionalProperties: false
    RegisterPayload:
      type: object
      required: [user]
      properties:
        user:
          $ref: '#/components/schemas/UserProfile'
      additionalProperties: false
    LoginPayload:
      type: object
      required: [user, auth]
      properties:
        user:
          $ref: '#/components/schemas/UserProfile'
        auth:
          $ref: '#/components/schemas/AuthPayload'
      additionalProperties: false
    RefreshPayload:
      type: object
      required: [auth]
      properties:
        auth:
          $ref: '#/components/schemas/AuthPayload'
      additionalProperties: false
    RegisterEnvelope:
      type: object
      required: [status, data, meta]
      properties:
        status:
          type: string
          const: ok
        data:
          $ref: '#/components/schemas/RegisterPayload'
        meta:
          $ref: '#/components/schemas/Meta'
      additionalProperties: false
    LoginEnvelope:
      type: object
      required: [status, data, meta]
      properties:
        status:
          type: string
          const: ok
        data:
          $ref: '#/components/schemas/LoginPayload'
        meta:
          $ref: '#/components/schemas/Meta'
      additionalProperties: false
    RefreshEnvelope:
      type: object
      required: [status, data, meta]
      properties:
        status:
          type: string
          const: ok
        data:
          $ref: '#/components/schemas/RefreshPayload'
        meta:
          $ref: '#/components/schemas/Meta'
      additionalProperties: false
    AcceptedEnvelope:
      type: object
      required: [status, data, meta]
      properties:
        status:
          type: string
          const: ok
        data:
          $ref: '#/components/schemas/AcceptedPayload'
        meta:
          $ref: '#/components/schemas/Meta'
      additionalProperties: false
...